# openapi.yaml
components:
  schemas:
...
    PaginatedBodyRequestTraceView:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/PaginatedBodyRequestTraceView.json
          format: uri
          readOnly: true
          type: string
        items:
          items:
            $ref: "#/components/schemas/RequestTraceView"
          nullable: true
          type: array
        pagination:
          $ref: "#/components/schemas/PaginationInfo"
      required:
        - items
        - pagination
      type: object
...
    PaginationInfo:
      additionalProperties: false
      properties:
        hasMore:
          type: boolean
        nextCursor:
          type: string
      required:
        - hasMore
      type: object
    PicoTeraError:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/PicoTeraError.json
          format: uri
          readOnly: true
          type: string
        code:
          type: string
        details:
          items:
            type: string
          nullable: true
          type: array
        message:
          type: string
      required:
        - message
        - code
        - details
      type: object
...
    RequestTraceView:
      additionalProperties: false
      properties:
        cacheReadTokens:
          format: int64
          type: integer
        cacheWrite1hTokens:
          format: int64
          type: integer
        cacheWriteTokens:
          format: int64
          type: integer
        firstRequestAt:
          type: string
        id:
          type: string
        inputTokens:
          format: int64
          type: integer
        lastRequestAt:
          type: string
        metaRequestCount:
          format: int64
          type: integer
        modelCosts:
          items:
            $ref: "#/components/schemas/TraceCostView"
          nullable: true
          type: array
        outputTokens:
          format: int64
          type: integer
        parentSpanId:
          type: string
        projectId:
          format: int32
          type: integer
        totalTokens:
          format: int64
          type: integer
        upstreamRequestCount:
          format: int64
          type: integer
        userMessagePreview:
          type: string
      required:
        - id
        - parentSpanId
        - metaRequestCount
        - upstreamRequestCount
        - totalTokens
        - inputTokens
        - cacheReadTokens
        - outputTokens
        - cacheWriteTokens
        - cacheWrite1hTokens
        - modelCosts
      type: object
...
    TraceCostView:
      additionalProperties: false
      properties:
        amount:
          format: double
          type: number
        currency:
          type: string
      required:
        - currency
        - amount
      type: object
...
info:
  title: PicoTera Management API
  version: 1.0.0
openapi: 3.0.3
paths:
...
  /api/picotera/request-traces:
    get:
      operationId: listRequestTraces
      parameters:
        - explode: false
          in: query
          name: limit
          schema:
            default: 20
            example: 20
            format: int32
            maximum: 100
            minimum: 1
            type: integer
        - explode: false
          in: query
          name: cursor
          schema:
            example: eyJpZCI6MX0=
            type: string
        - explode: false
          in: query
          name: startAt
          schema:
            type: string
        - explode: false
          in: query
          name: endAt
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedBodyRequestTraceView"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PicoTeraError"
          description: Error
      summary: List request traces
...